LU-14275 tests: add ior_CLEANUP
Sometimes we need to run read IOR test separately from
write test. IOR -r test requires file exists before test starts.
If ior_CLEANUP set to false $testdir is not removed and
$testdir/iorData file can be used for IOR -r further test.

Test-Parameters: trivial testlist=parallel-scale
HPE-bug-id: LUS-9560
Reviewed-on: https://review.whamcloud.com/41079
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>

+++ lustre/tests/functions.sh
@@ -611,6 +611,7 @@ run_ior() {
 	ior_xferSize=${ior_xferSize:-1M}
 	ior_type=${ior_type:-POSIX}
 	ior_DURATION=${ior_DURATION:-30}        # minutes
+	ior_CLEANUP=${ior_CLEANUP:-true}
 	local multiplier=1
 	case ${ior_blockUnit} in
 		[G])
@@ -692,7 +693,7 @@ run_ior() {
     if [ $rc != 0 ] ; then
         error "ior failed! $rc"
     fi
-    rm -rf $testdir
+    $ior_CLEANUP && rm -rf $testdir || true
 }
 
 run_mib() {
